Refined MiniMessage feature, UI updates, and cleanup
